BUG: Fix a bug in snapmerge due to writing to closed file stream
Program `snapmerge` had a bug: pointer FP has been used both for
checking that the input parameter file can be opened and for sending log
messages via `fprintf`.
FP was set to `stdout` later than the first `fprintf` is called.
It was leading to the undefined behavior: sometimes the program hanged
but sometimes worked.